Completion of Front end & Back end
Now front end can communicate with the back end.

May need to change the back end as it is really specific requirements to do things.

Also removed the middleware, found a way to do the job without it

The PUT and DELETE needed to have its own script for sending requests due to requirement for PUT and DELETE method